refactor: custom event source
