+++ packages/module-event-source/src/client/custom-event-sources/collectionCustomEventSources.ts
@@ -1,4 +1,4 @@
-export const collectionWorkflowsManager = {
+export const collectionCustomEventSources = {
   name: 'customEventSources',
   fields: [
     {
