+++ packages/module-event-source/src/client/custom-event-sources/useSyncCustomEventSource.ts
