+++ packages/module-event-source/src/locale/en-US.json
@@ -1,5 +1,6 @@
 {
   "Custom event source": "Custom event source",
   "Event source": "Event source",
+  "Sync": "Sync",
   "Test": "Test"
 }
