+++ packages/module-event-source/src/locale/zh-CN.json
@@ -2,6 +2,7 @@
   "Custom event source": "自定义事件源",
   "Event source": "事件源",
   "Resource": "资源",
+  "Sync": "同步",
   "Test": "测试",
   "TriggerOnAssociation": "关联时触发",
   "Type": "类型"
